Dev ecs (#29)
Successfully remodeled codebase to an ECS design. I removed the layerstack (an unnecessary component) and added several systems as well as components. I removed the node class and scene graph, so now all we need is the entity class.

I also removed the SFML binaries from the source tree and added SFML and JSON as vendor submodules in the project.

A lot of work went into the cmake build for this as well. Currently everything works.